卒業研究/data_viz_project/requirements.txt
pandas
matplotlib
requests
openai
python-dotenv
